RAG/requirements.txt
azure.identity
langchain-core
langchain-openai
langchain-community

langchain-chroma

pypdf

streamlit

protobuf==3.19.6
faiss-cpu